Move extra handlers logic to its own module
Also put the registering of extra handlers out of the controller
so we can make sure the auto reloader will not whipe it out.
